Update patch set 3
Patch Set 3: Looks good to me, but someone else must approve

My only concern with crediting is that author remain the way it was initially submitted as just zbwu. It is not about crediting my work. It is making sure ample credit is given for the work that was done before it even reached me. Had I known this needed to be uploaded, I would have provided a version that documents this better.

Patch-set: 3