Update patch set 5
Patch Set 5:

I do not follow your local changes to know where you keep it. I based my comment on what you hosted in your most recent change and the last known location of the files you submitted previously.
A fuss is not simply asking the original author to be substantially credited compared to minor "forward ported" changes. That is asking you to do what is right. I apologize for the inconvenience.

Patch-set: 5