msm7x30: Camera HAL for HTC Sense-based devices
The proprietary camera files must be from an HTC Sense build for this
HAL to function properly. These files include:

libcamera.so
libmmipl.so
libmmjpeg.so
liboemcamera.so
libcamera.so
libmmipl.so
libmmjpeg.so
liboemcamera.so

This code originated for Ace (zbwu86) and was later modified to add
support for front-facing camera (nico), all msm7x30 devices, reverse
front-facing camera, and inline overlay support (twisted).
In this condition both camera and front-facing camera are fully
functional, but the camcorder will only work with a third-party
application.