Update patch set 9
Patch Set 9:

I submitted an update to clean up the library list and explicitly reference the frameworks includes. As far as credit for the FFC, when you and him work out who is actually supposed to have their name, it will be fixed for the final revision, but for now that it cant wait until it works.

Patch-set: 9