Update patch set 10
Patch Set 10:

Something that's still not working is the mirroring of the ffc. The preview image should be mirrored but it's not. That's also why the preview image rotates in the wrong direction when you rotate the device. So it looks upside down when holding the device in landscape. But it's only because of missing mirror effect. I strongly believe the error is in liboverlay. In the logcat you first see an error when set source is called and then a lot of errors because the channel has not been set up. The first error happens when the ioctl for MSMFB_OVERLAY_SET is called. Would be cool if someone could fix this...

Patch-set: 10