Update patch set 10
Patch Set 10:

this is gonna get a monster thread :)

preview works fine from start with my preview fix applied to camera.

I do also have preview in face unlock! But maybe it's because I commented out the "return 0" in the set_parameters part of the cameraHal! That fixes face unlock but leads to the known instability of camera when ffc was active before a restart. Then the camera keeps crashing until you delete app data...

Patch-set: 10