msm7x30: Camera HAL for HTC Sense devices (1/2)
The proprietary camera files must be from an HTC Sense build for this
HAL to function properly. These files include:
libcamera.so
libmmipl.so
libmmjpeg.so
liboemcamera.so

This camera HAL was based on GB ported drivers from zbwu86 with FFC by
nico.
msm7x30 device support, face unlock, reverse FFC, and panorama added,
but the camcorder is not yet functioning.