Update patch set 10
Patch Set 10:

Interesting, because when I tested it, without the flag it was inverted and with the flag it was corrected. I am curious, though, as to why the results of so many devices seems different to what is seen by nexx and yourself. There has to be some external factor in your sources that is causing varied results.
I am also testing converting the pixel format to see if it yields better results, as most other HALs use RGB, but I am simply trying for a higher grade YUV

Patch-set: 10