Update patch set 12
Patch Set 12:

I am unsure if you are using the reverse FFC or not and which code is actually causing you the issues as it seems with each change you are having issues with a different version. Hopefully this is good enough to allow you to actually test it.

Patch-set: 12