Update patch set 2
Patch Set 2:

Ok, no problem, but for reference you did remove it in patch set 11:
http://review.cyanogenmod.com/#patch,sidebyside,12793,12,msm7x30.mk ;)

Looking forward to seeing this new HAL work, the CM9 Saga users are sure to be pleased if it works better than the existing one. Would appreciate it if you add me for review when it's uploaded.

Patch-set: 2