Update patch set 5
Patch Set 5:

Sorry guys, but this is just a first draft of the new code that was merged this morning. You are more than welcome to spend the entire day ripping it apart and commenting, but I won't be able to fix it up and post another revision for a day or two while everything is verified and cleaned. Gluck.

Patch-set: 5