Update patch set 7
Patch Set 7:

I lied. One more because all the issues mentioned were easily fixed by adding one chunk I forgot to port forward, and altering one method that was previously used. It currently has the issue with starting the camera in FFC, but I am not going to "hack" that so that a proper solution can be explored instead. Merry X-mas!

Patch-set: 7