msm7x30: Camera HAL for HTC Sense devices (1/2)
The proprietary camera files must be from an HTC Sense build for this
HAL to function properly. These files include:

libcamera.so
libmmipl.so
libmmjpeg.so
liboemcamera.so

This camera HAL was based on GB ported drivers from zbwu86 with FFC by
nico/nexx, and rewrite from QiSS Me.
Face unlock and panorama are supported, if available.

Required:
http://review.cyanogenmod.com/#change,13317
http://review.cyanogenmod.com/#change,13285

Optional for devices with preview issues:
http://review.cyanogenmod.com/#change,13564