Update patch set 10
Patch Set 10:

It was brought to my attention that the media_profiles currently available may cause issues with the camcorder and / or FFC, so the one I am using is https://github.com/ThePlayground/android_device_htc_msm7x30-common/blob/ics/system/etc/media_profiles.xml
If this is an issue that switching files resolves, please let me know so I can make that a part of the update. Thanks.

Patch-set: 10