Update patch set 14
Patch Set 14: I would prefer that you didn't submit this

The next revision is going to be uploaded as a new change to test if a hal built against updates that are not yet merged will still work with the current layout. Hopefully the new options will be properly ignored as there is no clean way to bring the camera application current that retains proper history yet.

Patch-set: 14