msm7x30: Camera HAL for HTC Sense devices (1/2)
The proprietary camera files must be from an HTC Sense build for this
HAL to function properly. These files include:

libcamera.so
libmmipl.so
libmmjpeg.so
liboemcamera.so

This camera HAL was based on GB ported drivers from zbwu86 with FFC by
nico/nexx, the rewrite from QiSS Me, and reconfigured portions of
rapmv78's qsd8k HAL layout.
Face unlock and panorama are supported, if available.

Required:
http://review.cyanogenmod.com/#change,13317
http://review.cyanogenmod.com/#change,13285

Optional (preview issues):
http://review.cyanogenmod.com/#change,13564

Notes:
CAF options are within an ifdef to allow use in testing of newer
parameters and for when officially integrated.
An alternate media_profiles is needed in those situations.