Update patch set 1
Patch Set 1:

By forcing the rotation and lower the resolution to 15 FPS, the recording will work but will not save. In newer camera updates, the resolution was to force a rotation in the HAL for ALL parameters and use the camera jni to override the rotation for the images. I am working on a solution that is a little less "hack"

Patch-set: 1