Update patch set 4
Patch Set 4:

Twisted, to fix the FFC crashing on setParameters, open and close the rear cam once. Currently I'm doing this on boot with an APK which listens for BootCompletedReceiver.

The APK opens then closes the rear cam, after that FFC works as expected and there's no more crash on Camera.apk after a reboot. Face unlock also works on boot too.

Source for the APK is here:
https://github.com/nexx/android_packages_apps_FFCFix

Patch-set: 4