Update patch set 5
Patch Set 5:

Twisted, Nico and I have been testing the camera HAL today combined with the open-source QCom OMX updates. We now have video recording working though there are a couple of issues in the code which you could perhaps explain.

1) *preferred_rate = "15";
This breaks recording from the rear cam, setting it to 30, or locking the FPS to 15 in media_profiles.xml fixes that.

2) info->orientation = 90;
This breaks FFC as the image is now 90 degrees CW instead of 270 degrees CCW. Everything is upside down :P

Setting info->orientation to anything else breaks recording with no useful output as to why, even setting it to 90 for the rear-cam only causes a failure in recording.

The FFC records fine regardless.

Patch-set: 5