Update patch set 5
Patch Set 5:

The locked framerate is during testing of some other parameters to ensure camera works, it will then be updated.
The rotation issue is the camera application. The rotation of 90 is the default and has been since Gingerbread. Removing it will disable the video camera functionality for most devices as stated numerous times. The resolution is to adjust your camera application to account for it.

Patch-set: 5