Update patch set 5
Patch Set 5:

The rotation is not only the camera app. Face-unlock and 3rd party camera apps on FFC are also upside down when your code forces the orientation to 90.

As for locking the framerate, recording from the rear camera doesn't work *full-stop* with it forced to 15. Either setting it to 15 in media_profiles.xml or forcing 30fps instead makes it work as expected.

Patch-set: 5