Update patch set 5
Patch Set 5:

To be more productive than the last string of changes that led to finishing where they started, feel free to submit actual suggested changes that fix the code for your device but do not interfere with the current code (framerate excluded).
This would be far more useful than the current method of simply stating lines that are suspected issues, but happen to be required by all but your device. I will be glad to make adjustments, but not at the risk of disabling 5 devices for one to work.

Patch-set: 5