Update patch set 6
Patch Set 6:

Please note the new flag "TARGET_INVERSE_ROTATE" that will set the rotation to 0. This should be set in the local board configuration of devices with rotation issues at the default 90. If this does not correct the rotation, please provide a value that would, as it was noted before that simply letting rotation set itself had issues as well.

Patch-set: 6