Update patch set 6
Patch Set 6:

The mytouch 4g and thunderbolt are the primary ones being tested against and no, they do not have the issues you seem to. Keep in mind, you are not accounting for the fact that video recording occurs primarily from the back camera, and that the flag for reverse FFC is available. Either way, I have been making guesses as to what it is exactly that will resolve the issues as I am only getting a user-level "error report" in terms of feedback. Hopefully this patch resolves it.

Patch-set: 6