Update patch set 8
Patch Set 8:

The alternative would be a call to if facing == front or something of that nature, but this fit into the existing ifdef a little easier.

Massive cleanup is needed for the whitespace. This is relocated in my local source, so copying caused an excessive amount of tabs that will be fixed.

Patch-set: 8