Update patch set 9
Patch Set 9:

Thanks Sean, that works.

As for not posting fixes Twisted. I don't see why I should bother when, once again, you're removed the credits for the code I wrote into the HAL. All the initial FFC work was done by Nico and myself and you've been reminded about this in the past.

I don't know why you seem to have an issue with me, but it's really starting to wear a bit thin =(

Patch-set: 9