Update patch set 9
Patch Set 9:

I was hoping to use something like that, but wasn't sure if it was covered in the scope of the function. I was sticking to the same variables already present. I am submitting it with the facing method, but worst case I can convert to that instead.

Patch-set: 9